PR feedback; moved time formatting function, reverted JSX change
